AccountManagerPlugin: Improve password file handling, closes #4276 and #4897.
htpasswd.bak files are not deleted anymore when updating htpasswd file
in the same directory and it's ownership is preserved as well, preventing
a DoS by inaccessible user file in shared use.
Furthermore we use universal newline support, if build-in, so line endings
are always seen as '\n' by our code regardless of real formatting
(Macintosh convention '\r', Windows convention '\r\n').  However the
actual end-of-line style is probed and preserved on file updates.
Finally the file is not left open for an indeterminated amount of time
after file access, what wasn't reported but deduced by own code studies
to be a potential issue and therefor fixed as a precaution.

git-svn-id: https://trac-hacks.org/svn/accountmanagerplugin/trunk@9272 7322e99d-02ea-0310-aa39-e9a107903beb