AccountManagerPlugin: Consolidate plugin permissions, refs #7700.
Central definition in the `AccountManager` module is not only more
straight-forward, now actions are guaranteed to exist even with admin panels
disabled.

As a side-effect this fixes irritating `ACCTMGR_USER_ADMIN` duplicate entries
in the permission select box as seen i.e. in Trac 0.11.

Added unit test to ensure this works expectedly, and added a more intuitive
inheritance of `EMAIL_VIEW` by `ACCTMGR_USER_ADMIN` too, that was missing
because of wrong syntax in `AccountManagerAdminPanels.get_permission_actions`
before.

git-svn-id: https://trac-hacks.org/svn/accountmanagerplugin/trunk@11927 7322e99d-02ea-0310-aa39-e9a107903beb